Initial WIP of 0.0.3
