--- api/context_keys.go
@@ -1,6 +1,0 @@
-package api
-
-const (
-	responseKey = "response"
-	errorKey    = "error"
-)
