--- api/response.go
@@ -1,6 +1,0 @@
-package api
-
-type APIResponse struct {
-	Data  interface{} `json:"data,omitempty"`
-	Error string      `json:"error,omitempty"`
-}
