+++ src/cmd/cistern/version.go
@@ -1,0 +1,3 @@
+package main
+
+const version = "0.0.3"
