+++ src/internal/net/snmp/snmp.go
@@ -1,0 +1,3 @@
+package snmp
+
+// TODO
