--- version.go
@@ -1,3 +1,0 @@
-package main
-
-const version = "0.0.2"
